Implement page-per-dataclass with field linking.
The reflection infrastrucure introduced to render the interpreter
provider page and parse the lift manifest is expanded to handle full
cross-referencing of @dataclass types as well as handling generating a
seperate page per dataclass. This is used to restructure `providers.md`
and fill out the skeleton of the full lift `manifest.md` doc tree.

Added as well are a few facilities to ensure documentation is present
and not broken:
+ Invalid cross-references produce CI errors during sphinx html gen.
+ Invalid external links produce CI errors during sphinx linkcheck.